flask/requirements.txt
flask
Flask-SQLAlchemy
flask-login
python-whois